# src/server/routes/admin.routes.js
import _ from 'lodash';
import {Router} from 'express';
import {    
    adminLoginHandler,
    adminSignupHandler,
    deleteAdminHandler,
    addNewAdminHandler,
    getAdminDetailsHandler,
    getAdminListHandler,
    updateAdminDetailsHandler,
    getDashboardData,
    contactusHandler
} from '../../common/lib/admin/adminHandler';
import responseStatus from "../../common/constants/responseStatus.json";
import responseData from "../../common/constants/responseData.json";
import protectRoutes from "../../common/util/protectRoutes";

const router = new Router();

router.route('/signup').post(async (req, res) => {
    try {
        if (!_.isEmpty(req.body)) {
            const outputResult = await adminSignupHandler(req.body);
            res.status(responseStatus.STATUS_SUCCESS_OK);
            res.send({
                status: responseData.SUCCESS,
                data: {
                    admin: outputResult.admin,
                    token: outputResult.token
                }
            });
        } else {
            throw 'No request body sent';
        }
    } catch (err) {
        console.log(err);
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err }
        });
    }
});

router.route('/contact-us').post(async (req, res) => {
    try {
        if (!_.isEmpty(req.body)) {
            const send_mail = await contactusHandler(req.body);
            res.status(responseStatus.STATUS_SUCCESS_OK).json({
                status: responseData.SUCCESS,
                data: { send_mail }
            });
        } else {
            throw "Failed Sending Contact-Us Mail"
        }
    } catch (err) {
        console.error("Error in router:", err);
        res.status(responseStatus.INTERNAL_SERVER_ERROR).json({
            status: responseData.ERROR,
            data: { message: err.message || err}
        });
    }
});

router.route("/login").post(async (req, res) => {
    try {
      if (!_.isEmpty(req.body)) {
        const adminData = await adminLoginHandler(req.body);
        res.status(responseStatus.STATUS_SUCCESS_OK).json({
          status: responseData.SUCCESS,
          data: { admin: adminData },
        });
      } else {
        throw "No credentials provided";
      }
    } catch (err) {
      console.error(err);
      res.status(responseStatus.INTERNAL_SERVER_ERROR).json({
        status: responseData.ERROR,
        data: { message: err },
      });
    }
  });

router.route('/list').post(protectRoutes.authenticateToken,async (req, res) => {
    try {
      let filter = {};
      filter.query = {};
  
      const inputData = { ...req.body };
      if (inputData) {
        filter.pageNum = inputData.pageNum ? inputData.pageNum : 1;
        filter.pageSize = inputData.pageSize ? inputData.pageSize : 50;
  
        if (inputData.filters) {
          filter.query = inputData.filters;
        }
      } else {
        filter.pageNum = 1;
        filter.pageSize = 50;
      }
  
      filter.query = { ...filter.query };
  
      const outputResult = await getAdminListHandler(filter);
      res.status(responseStatus.STATUS_SUCCESS_OK);
      res.send({
        status: responseData.SUCCESS,
        data: {
          adminList: outputResult.list ? outputResult.list : [],
          adminCount: outputResult.count ? outputResult.count : 0,
        },
      });
    } catch (err) {
      console.log(err);
      res.status(responseStatus.INTERNAL_SERVER_ERROR);
      res.send({
        status: responseData.ERROR,
        data: { message: err },
      });
    }
  });

router.route('/dashboard').get(protectRoutes.verifyAdmin,async(req,res)=>{
    try{
        let filter = {};
        filter.query = {};

        const inputData = { ...req.body };
        if (inputData) {
            filter.pageNum = inputData.pageNum ? inputData.pageNum : 1;
            filter.pageSize = inputData.pageSize ? inputData.pageSize : 50;

            if (inputData.filters) {
            filter.query = inputData.filters;
            }
        } else {
            filter.pageNum = 1;
            filter.pageSize = 50;
        }

        filter.query = { ...filter.query };
  
        const dashboard_data = await getDashboardData(filter);
        res.status(responseStatus.STATUS_SUCCESS_OK);
        res.send({
            status: responseData.SUCCESS,
            data: { dashboard_data }
        });
    }
    catch(err)
    {
        console.log(err);
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err },
        });
    }
});

router.route('/new').post(async (req, res) => {
    try {
       if (!_.isEmpty(req.body)) {
            const outputResult = await addNewAdminHandler(req.body.admin);
            res.status(responseStatus.STATUS_SUCCESS_OK);
            res.send({
                status: responseData.SUCCESS,
                data: {
                    admin: outputResult ? outputResult : {}
                }
            });
        } else {
            throw 'no request body sent'
        }
    } catch (err) {
        console.log(err)
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err }
        });
    }
});

router.route('/:id').get(async (req, res) => {
    try {
        if (req.params.id) {
            const gotAdmin = await getAdminDetailsHandler(req.params);
            res.status(responseStatus.STATUS_SUCCESS_OK);
            res.send({
                status: responseData.SUCCESS,
                data: {
                    admin: gotAdmin ? gotAdmin : {}
                }
            });
        } else {
            throw 'no id param sent'
        }
    } catch (err) {
        console.log(err)
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err }
        });
    }
});

router.route('/:id/update').post( async (req, res) => {
    try {
        if (!_.isEmpty(req.params.id) && !_.isEmpty(req.body) && !_.isEmpty(req.body.admin)) {
            let input = {
                objectId: req.params.id,
                updateObject: req.body.admin
            }
            const updateObjectResult = await updateAdminDetailsHandler(input);
            res.status(responseStatus.STATUS_SUCCESS_OK);
                res.send({
                    status: responseData.SUCCESS,
                    data: {
                        admin: updateObjectResult ? updateObjectResult : {}
                    }
                });
        } else {
            throw 'no body or id param sent'
        }
    } catch (err) {
        console.log(err)
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err }
        });
    }
});

router.route('/:id/remove').post(async(req, res) => {
    try {
        if (req.params.id) {
            const deletedAdmin = await deleteAdminHandler(req.params.id);
            res.status(responseStatus.STATUS_SUCCESS_OK);
            res.send({
                status: responseData.SUCCESS,
                data: {
                    hasAdminDeleted: true
                }
            });
        } else {
            throw 'no id param sent'
        }
    } catch (err) {
        console.log(err)
        res.status(responseStatus.INTERNAL_SERVER_ERROR);
        res.send({
            status: responseData.ERROR,
            data: { message: err }
        });
    }
});

export default router;
  